% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.datetime.r
\name{as.Date0}
\alias{DOW}
\alias{DateAndHEAsPOSIX}
\alias{DateAsDateTime}
\alias{DateFromDateTime}
\alias{FirstDayOfMonth}
\alias{HourFromDateTime}
\alias{MonthFromDate}
\alias{YearFromDate}
\alias{as.Date0}
\title{Format Date-Time}
\usage{
as.Date0(date)

MonthFromDate(date)

FirstDayOfMonth(date)

YearFromDate(date)

DateAsDateTime(date)

DOW(date)

HourFromDateTime(datetime)

DateFromDateTime(datetime)

DateAndHEAsPOSIX(date, HE)
}
\arguments{
\item{date}{Date in either 'character' or 'date' format.}

\item{datetime}{date-time in 'POSIXct' format.}

\item{HE}{Integer (1:24) indicating hour-ending.}
}
\description{
Formatting function for date-time data.
}
\details{
DateFromDateTime is different from simply using as.Date(date-time), because date-time is stored in
standard time with time zone in R.
}
\section{Functions}{
\itemize{
\item \code{MonthFromDate}: Given a date, rerurn the month as numeric

\item \code{FirstDayOfMonth}: Given a date, returns the first day of the month

\item \code{YearFromDate}: Given a date, rerurn the year as numeric

\item \code{DateAsDateTime}: Given a date, returns 12:00am of the date in POSIXct format

\item \code{DOW}: Day of week from a date

\item \code{HourFromDateTime}: Given a date-time, return the hour as numeric

\item \code{DateFromDateTime}: Given a date-time, returns a date.

\item \code{DateAndHEAsPOSIX}: Given a date and HE, returns the date-time for the beginning of the hour.
}}
\examples{
as.Date0("1/1/2000")
as.Date1("1/1/2000"); as.Date1("2000-1-1")

dates <- seq(as.Date("2000-1-1"), length.out = 6, by="mon")+15
dates
MonthFromDate(dates)
FirstDayOfMonth(dates)
YearFromDate(dates)
DOW(dates)
DateAndHEAsPOSIX(dates, c(1:10))

datetimes <- DateAsDateTime(dates)+8*60*60
datetimes
HourFromDateTime(datetimes)
DateFromDateTime(datetimes)

}

