% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcDailyPeakPeriods.r
\name{CalcDailyPeakPeriods}
\alias{CalcDailyPeakPeriods}
\title{Convert hourly data to HLH/LLH Daily}
\usage{
CalcDailyPeakPeriods(date, mtx, market = "WECC", colname = "")
}
\arguments{
\item{date}{Consecutive daily date data in 'date' format.}

\item{mtx}{A matrix with 24 columns corresponding to each hour. Number of rows need to match the length of date.}
}
\description{
Given 24 hour matrix, it returns HLH/LLH Daily data frame.
}
\details{
WECC definition was used in HLH/LLH split.
}
\examples{
dates <- seq(as.Date("2000-1-1"), length.out = 20, by="day")
mtx <- matrix(1:(24*20), ncol=24)
CalcDailyPeakPeriods(dates, mtx, colname="Load_aMW_")

}

