% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.table.r
\name{GetDateTable}
\alias{GetDateTable}
\title{Get Date Table}
\usage{
GetDateTable(dates, col.names = c("Date", "Month", "Year", "DOW", "Peak",
  "pk.nHours", "opk.nHours", "pk.weight", "opk.weight"))
}
\arguments{
\item{dates}{an array of date objects}

\item{col.names}{A list of columns included in the returned table.
Default returns all columns: c("Date","Month","Year", "DOW", "Peak", "pk.nHours","opk.nHours", "pk.weight", "opk.weight")}
}
\description{
Given dates, it returns a data frame with peak and off-peak information.
}
\examples{
dates <- seq(as.Date("2015-1-1"), length.out = 10, by=1)
GetDateTable(dates)

# Calculate number of on-peak and off-peak hours
dates <- seq(as.Date("2015-1-1"),as.Date("2020-1-1"), by=1)
date.df <- GetDateTable(dates, c("Date", "pk.nHours", "opk.nHours"))
aggregate(date.df[,-1], by=list("Date"=FirstDayOfMonth(date.df$Date)), sum)
}

